actions/copro_commitment_link.go
package actions

import (
	"database/sql"
	"net/http"

	"github.com/Iledant/PreLoRUGo/models"
	"github.com/kataras/iris"
)

// LinkCommitmentsCopros handles the post request of a batch of CoproCommiment
func LinkCommitmentsCopros(ctx iris.Context) {
	var l models.CoproCommitmentBach
	if err := ctx.ReadJSON(&l); err != nil {
		ctx.StatusCode(http.StatusInternalServerError)
		ctx.JSON(jsonError{"Liens engagements copros, décodage : " + err.Error()})
		return
	}
	db := ctx.Values().Get("db").(*sql.DB)
	if err := l.Save(db); err != nil {
		ctx.StatusCode(http.StatusInternalServerError)
		ctx.JSON(jsonError{"Liens engagements copros, requête : " + err.Error()})
		return
	}
	ctx.StatusCode(http.StatusOK)
	ctx.JSON(jsonMessage{"Liens engagements copros importés"})

}
